#375 Remove parent verification in .equals/.hashCode and remove last usages of HashSet<DiagramElement>.
Parent verification for FieldNodes and CallNodes is not required because it is not possible to change the parent of these Nodes through the application interface. Since there are no longer any cycle concerns in the implementation of .equals/.hashCode for nodes, it is now possible to elevate the check of children equality to the AbstractNode level.

It was then discovered that some instances of HashSet<DiagramElement> were remaining in different parts of the code such as:
- DiagramBuilder
- ControlFlow
- DiagramCanvasController
- TestSequenceDiagramBuilder

The usage of HashSet is problematic as it relies on .equals (which has been modified and is no longer identical to ==). It has been replaced with the usage of IdentityHashMap (which relies on ==) with default values of -1. A further refactor could be performed to implement a IdentityHashSet, or to use an external library providing this functionality.