Add a performance test to measure execution time degradations.
Lines must be uncommented to execute the performance test. It will always "pass". The real information lies in the printed average execution times.

The tests consist of translating nodes and getting their new bounds from NodeViewerRegistry. Currently, for each test, there are 10 rounds of 10000 iterations each. There is a test for 1 node, one for 3 nodes, and one for 10 nodes. The time is reported in milliseconds.